refactor: remove FX from breadth workflows
