+++ backend/tests/unit/test_breadth_universe.py
@@ -1,14 +1,11 @@
 from datetime import date
 
 import pandas as pd
-import pytest
 from app.services.breadth.market_policy import get_breadth_market_policy
 from app.services.breadth.types import BreadthFormulaPolicy, BreadthUniverseMember
 from app.services.breadth.universe import (
-    MissingHistoricalFXError,
     build_breadth_universe_snapshots,
     classify_metric_eligibility,
-    resolve_historical_fx_series,
     stockbee_eligibility_signature,
 )
 from app.services.point_in_time_universe_service import (
@@ -16,56 +13,6 @@ from app.services.point_in_time_universe_service import (
     PointInTimeUniverseMember,
     hash_point_in_time_universe_symbols,
 )
-
-
-def test_historical_fx_prefers_exact_date_then_prior_within_seven_days():
-    requested = (date(2026, 8, 20), date(2026, 8, 21))
-    result = resolve_historical_fx_series(
-        "HKD",
-        requested,
-        {
-            date(2026, 8, 14): 0.127,
-            date(2026, 8, 21): 0.128,
-        },
-        max_age_days=7,
-    )
-
-    assert result.loc[pd.Timestamp("2026-08-20")] == pytest.approx(0.127)
-    assert result.loc[pd.Timestamp("2026-08-21")] == pytest.approx(0.128)
-
-
-def test_historical_fx_rejects_quote_older_than_seven_calendar_days():
-    with pytest.raises(MissingHistoricalFXError) as exc_info:
-        resolve_historical_fx_series(
-            "HKD",
-            (date(2026, 8, 22),),
-            {date(2026, 8, 14): 0.127},
-            max_age_days=7,
-        )
-
-    assert exc_info.value.currency == "HKD"
-    assert exc_info.value.calculation_date == date(2026, 8, 22)
-
-
-def test_historical_fx_never_uses_future_quote():
-    with pytest.raises(MissingHistoricalFXError):
-        resolve_historical_fx_series(
-            "HKD",
-            (date(2026, 8, 21),),
-            {date(2026, 8, 22): 0.128},
-            max_age_days=7,
-        )
-
-
-def test_historical_fx_usd_is_identity_without_observations():
-    result = resolve_historical_fx_series(
-        "usd",
-        (date(2026, 8, 20), date(2026, 8, 21)),
-        {},
-        max_age_days=7,
-    )
-
-    assert result.tolist() == [1.0, 1.0]
 
 
 class _UniverseResolver:
